+++ src/application/structs/UnlockStatus.h
@@ -1,0 +1,32 @@
+﻿#pragma once
+
+struct UnlockStatus
+{
+    int currentEmblems;
+    int emblemGoal;
+
+    bool sonicUnlocked;
+    bool sonicLightShoesUnlocked;
+    bool sonicCrystalRingUnlocked;
+    bool sonicAncientLightUnlocked;
+
+    bool tailsUnlocked;
+    bool tailsJetAnkletUnlocked;
+    bool tailsRhythmBadgeUnlocked;
+
+    bool knucklesUnlocked;
+    bool knucklesShovelClawUnlocked;
+    bool knucklesFightingGlovesUnlocked;
+
+    bool amyUnlocked;
+    bool amyLongHammerUnlocked;
+    bool amyWarriorFeatherUnlocked;
+
+    bool gammaUnlocked;
+    bool gammaLaserBlasterUnlocked;
+    bool gammaJetBoosterUnlocked;
+
+    bool bigUnlocked;
+    bool bigLifeRingUnlocked;
+    bool bigPowerRodUnlocked;
+};