+++ src/output/itemRepository/ItemRepository.h
@@ -1,6 +1,8 @@
 ﻿#pragma once
 #include <map>
 #include "../../application/structs/ItemData.h"
+#include "../../application/structs/UnlockStatus.h"
+
 
 
 class ItemRepository
@@ -11,9 +13,13 @@ public:
     ItemData GetItem(int itemId);
     std::map<int64_t, ItemData> GetItems();
     int AddEmblem();
+    int GetEmblemGoal() const;
+    void SetEmblemGoal(int emblemGoal);
     int GetEmblemCount();
+    UnlockStatus GetUnlockStatus();
 
 private:
     std::map<int64_t, ItemData> _itemData;
     int _emblemCount = 0;
+    int _emblemGoal = -1; 
 };